+++ tests/src/PredictionService.Tests/FeatureNormalizerTests.cs
@@ -1,6 +1,8 @@
 using PredictionService.Models;
 using PredictionService.ML;
 using Xunit;
+using Microsoft.Extensions.Logging;
+
 
 namespace PredictionService.Tests.ML;
 
@@ -318,7 +320,7 @@ public class FeatureNormalizerTests
             TenureDays = 730,
             ProductDiversity = 8,
             CategoryDiversity = 5,
-            AvgProductsPerOrder = 2.5f,
+            AvgProductsPerOrder = (decimal)2.5f,
             ReturnCount = 2,
             ReturnRate = 0.05m,
             TotalRefunded = 300,
